修复 Integer转Long报错:class java.lang.Integer cannot be cast to class java.lang.Long (java.lang.Integer and java.lang.Long are in module java.base of loader 'bootstrap')

+++ src/main/java/github/ag777/util/lang/ObjectUtils.java
@@ -158,7 +158,7 @@ public class ObjectUtils {
         if(obj instanceof Long) {
             return (Long) obj;
         } else if(obj instanceof Integer) {
-            return (long) obj;
+            return ((Integer) obj).longValue();
         } else if (obj instanceof Float) {
             return ((Float)obj).longValue();
         } else if (obj instanceof Double) {
